requirements.txt
argparse
numpy
scipy
scikit-learn
torchvision
ema_pytorch
pywavelets
matplotlib
mne
pandas
pyedflib
PyWavelets
tqdm
